moves subscriber attributes
