+++ feature/subscriber-attributes/src/main/java/com/revenuecat/purchases/subscriberattributes/caching/SubscriberAttributesCache.kt
@@ -1,0 +1,209 @@
+//  Purchases
+//
+//  Copyright © 2019 RevenueCat, Inc. All rights reserved.
+//
+
+package com.revenuecat.purchases.common.caching
+
+import android.content.SharedPreferences
+import com.revenuecat.purchases.common.debugLog
+import com.revenuecat.purchases.subscriberattributes.SubscriberAttribute
+import com.revenuecat.purchases.subscriberattributes.buildLegacySubscriberAttributes
+import com.revenuecat.purchases.subscriberattributes.buildSubscriberAttributesMapPerUser
+import org.json.JSONException
+import org.json.JSONObject
+
+private const val SHARED_PREFERENCES_PREFIX = "com.revenuecat.purchases."
+
+internal typealias AppUserID = String
+internal typealias SubscriberAttributeMap = Map<String, SubscriberAttribute>
+internal typealias SubscriberAttributesPerAppUserIDMap = Map<AppUserID, SubscriberAttributeMap>
+
+class SubscriberAttributesCache(
+    private val preferences: SharedPreferences,
+    private val apiKey: String
+) {
+    val attributionCacheKey = "$SHARED_PREFERENCES_PREFIX.attribution"
+    val subscriberAttributesCacheKey: String by lazy { "$SHARED_PREFERENCES_PREFIX$apiKey.subscriberAttributes" }
+
+    // region subscriber attributes
+
+    @Synchronized
+    fun setAttributes(appUserID: AppUserID, attributesToBeSet: SubscriberAttributeMap) {
+        val currentlyStoredAttributesForAllUsers = getAllStoredSubscriberAttributes()
+        val currentlyStoredAttributesForAppUserID =
+            currentlyStoredAttributesForAllUsers[appUserID] ?: emptyMap()
+        val updatedAttributesForUser = currentlyStoredAttributesForAppUserID + attributesToBeSet
+        val updatedAttributesForAllUsers =
+            currentlyStoredAttributesForAllUsers + mapOf(appUserID to updatedAttributesForUser)
+
+        preferences.edit().putAttributes(updatedAttributesForAllUsers).apply()
+    }
+
+    @Synchronized
+    fun getAllStoredSubscriberAttributes(): SubscriberAttributesPerAppUserIDMap =
+        preferences.getJSONObjectOrNull(subscriberAttributesCacheKey)
+            ?.buildSubscriberAttributesMapPerUser() ?: emptyMap()
+
+    @Synchronized
+    fun getAllStoredSubscriberAttributes(appUserID: AppUserID): SubscriberAttributeMap =
+        getAllStoredSubscriberAttributes()[appUserID] ?: emptyMap()
+
+    @Synchronized
+    fun getUnsyncedSubscriberAttributes(): SubscriberAttributesPerAppUserIDMap =
+        getAllStoredSubscriberAttributes()
+            .mapValues { (appUserID, attributesForUser) ->
+                attributesForUser.filterUnsynced(appUserID)
+            }.filterValues { it.isNotEmpty() }
+
+    @Synchronized
+    fun getUnsyncedSubscriberAttributes(appUserID: String) =
+        getAllStoredSubscriberAttributes(appUserID).filterUnsynced(appUserID)
+
+    private fun SubscriberAttributeMap.filterUnsynced(appUserID: AppUserID): SubscriberAttributeMap =
+        this.filterValues { !it.isSynced }
+            .also { unsyncedAttributesByKey ->
+                debugLog("Found ${unsyncedAttributesByKey.count()} unsynced attributes for appUserID: $appUserID \n" +
+                    if (unsyncedAttributesByKey.isNotEmpty()) {
+                        unsyncedAttributesByKey.values.joinToString("\n")
+                    } else {
+                        ""
+                    }
+                )
+            }
+
+    @Synchronized
+    fun clearSubscriberAttributesIfSyncedForSubscriber(appUserID: String) {
+        val unsyncedSubscriberAttributes = getUnsyncedSubscriberAttributes(appUserID)
+        if (unsyncedSubscriberAttributes.isNotEmpty()) {
+            return
+        }
+        debugLog("Deleting subscriber attributes for $appUserID from cache.")
+        val allStoredSubscriberAttributes = getAllStoredSubscriberAttributes()
+        val updatedStoredSubscriberAttributes =
+            allStoredSubscriberAttributes.toMutableMap().also {
+                it.remove(appUserID)
+            }.toMap()
+
+        preferences.edit().putAttributes(updatedStoredSubscriberAttributes).apply()
+    }
+
+    @Synchronized
+    fun cleanUpSubscriberAttributeCache(currentAppUserID: String) {
+        migrateSubscriberAttributesIfNeeded()
+        deleteSyncedSubscriberAttributesForOtherUsers(currentAppUserID)
+    }
+
+    @Synchronized
+    private fun deleteSyncedSubscriberAttributesForOtherUsers(currentAppUserID: String) {
+        debugLog("Deleting old synced subscriber attributes that don't belong to $currentAppUserID.")
+
+        val allStoredSubscriberAttributes = getAllStoredSubscriberAttributes()
+
+        val filteredMap =
+            allStoredSubscriberAttributes.map { (appUserID, attributesMap) ->
+                if (currentAppUserID != appUserID) {
+                    appUserID to attributesMap.filterValues { !it.isSynced }
+                } else {
+                    appUserID to attributesMap
+                }
+            }.toMap().filterValues { it.isNotEmpty() }
+
+        preferences.edit().putAttributes(filteredMap).apply()
+    }
+
+    private fun SharedPreferences.Editor.putAttributes(
+        updatedSubscriberAttributesForAll: SubscriberAttributesPerAppUserIDMap
+    ): SharedPreferences.Editor {
+        return this.putString(
+            subscriberAttributesCacheKey,
+            updatedSubscriberAttributesForAll.toJSONObject().toString()
+        )
+    }
+
+    private fun Map<AppUserID, SubscriberAttributeMap>.toJSONObject(): JSONObject {
+        val attributesObject = JSONObject().also { attributesJSONObject ->
+            this.forEach { (appUserID, subscriberAttributeMap) ->
+                val jsonObject = JSONObject().also { userJSONObject ->
+                    subscriberAttributeMap.forEach { (key, subscriberAttribute) ->
+                        userJSONObject.put(key, subscriberAttribute.toJSONObject())
+                    }
+                }
+                attributesJSONObject.put(appUserID, jsonObject)
+            }
+        }
+        return JSONObject().also { it.put("attributes", attributesObject) }
+    }
+
+    // endregion
+
+    // region migrate subscriber attributes
+
+    @Synchronized
+    private fun migrateSubscriberAttributesIfNeeded() {
+        getAllLegacyStoredSubscriberAttributes()
+            .takeIf { it.isNotEmpty() }
+            ?.let { legacySubscriberAttributes ->
+                migrateSubscriberAttributes(legacySubscriberAttributes)
+            }
+    }
+
+    @Synchronized
+    private fun migrateSubscriberAttributes(
+        legacySubscriberAttributesForAppUserID: SubscriberAttributesPerAppUserIDMap
+    ) {
+        val editor = preferences.edit()
+        val storedSubscriberAttributesForAll: SubscriberAttributesPerAppUserIDMap =
+            getAllStoredSubscriberAttributes()
+
+        val updatedStoredSubscriberAttributesForAll: MutableMap<String, SubscriberAttributeMap> =
+            storedSubscriberAttributesForAll.toMutableMap()
+
+        legacySubscriberAttributesForAppUserID.forEach { (appUserID: AppUserID, legacy: SubscriberAttributeMap) ->
+            val current: SubscriberAttributeMap = storedSubscriberAttributesForAll[appUserID] ?: emptyMap()
+            val updated: SubscriberAttributeMap = legacy + current
+            updatedStoredSubscriberAttributesForAll[appUserID] = updated
+            editor.remove(legacySubscriberAttributesCacheKey(appUserID))
+        }
+
+        editor.putAttributes(updatedStoredSubscriberAttributesForAll).apply()
+    }
+
+    // endregion
+
+    // region legacy subscriber attributes
+
+    fun legacySubscriberAttributesCacheKey(appUserID: String) =
+        "$subscriberAttributesCacheKey.$appUserID"
+
+    @Synchronized
+    private fun getAllLegacyStoredSubscriberAttributes(): SubscriberAttributesPerAppUserIDMap {
+        val legacySubscriberAttributesCacheKeyPrefix = legacySubscriberAttributesCacheKey("")
+        val allSubscriberAttributesKeys = preferences.all
+            ?.filterKeys { it.startsWith(legacySubscriberAttributesCacheKeyPrefix) }
+            ?.keys ?: emptySet()
+
+        return allSubscriberAttributesKeys.map { preferencesKey ->
+            val appUserIDFromKey: AppUserID =
+                preferencesKey.split(legacySubscriberAttributesCacheKeyPrefix)[1]
+            val subscriberAttributeMap: SubscriberAttributeMap =
+                preferences.getJSONObjectOrNull(preferencesKey)
+                    ?.buildLegacySubscriberAttributes() ?: emptyMap()
+            appUserIDFromKey to subscriberAttributeMap
+        }.toMap()
+    }
+
+    // endregion
+
+    private fun SharedPreferences.getJSONObjectOrNull(key: String): JSONObject? {
+        return this.getString(key, null)?.let { json ->
+            try {
+                JSONObject(json)
+            } catch (e: JSONException) {
+                null
+            }
+        }
+    }
+
+    // endregion
+}
